Solve Day 2
I'm not sure if it's because I spent some morning time on it and I'm
feeling slightly better or if the problem was easier but, I had a much
easier go on this problem. I thought about creating a struct to
represent the hand signs but there are so few permutations that pattern
matching seemed more pragmatic and I think I'm right. Luckily, they
didn't bust out Spock or Lizard on part two.