Solve Day 3
I was hoping it would be a little easier to use intersections but
HashSet was it. It was pretty straight forward for the first part but
the second part resulted in passing pointers to pointers. I'm not quite
sure why it was different but dereferencing it twice worked so I stuck
with it.

It was pretty tricky reversing the values of uppers and lowers but it
was still straight-forward enough to use the ascii values.