Solve Day 7
So this one took me a few days. Identifying the tree structure was
pretty easy but between being tired and it being my first time with
Structs in Rust I took it slow. There is some repetitive logic when
building the tree and I took some short cuts to avoid borrowing the
nodes but overall I'm happy with the solution.

It definitely feels like my Rust is getting better. Maybe it just feels
that way because autoformatting actually looks good.